fix(tg): narrow reply message_id to i32 for outbox reply_to (#1230)
The rich-first `reply` action (#1230) passed get_id()'s i64 straight
into send_markdown_outbox's reply_to: Option<i32> — E0308 blocked
every tip from compiling (cycle-18 red, run 32999957533). Cast to i32
matching the file's existing convention (edit/delete/pin sites).

Session-Id: d18ce16a-75a0-447c-90c7-ab7dabce4411

+++ src/brain/tools/telegram_send.rs
@@ -548,7 +548,7 @@ impl TelegramSendTool {
             &text,
             "tool",
             "reply",
-            Some(message_id),
+            Some(message_id as i32),
         )
         .await
         {
